apply clippy warnings fixes

+++ src/interpreter/stdlib/math/power.rs
@@ -15,7 +15,7 @@ pub fn std_pow(_: &mut Evaluator, args: Vec<Value>) -> Result<Value, Error> {
     match (base, exponent) {
         (Value::Integer(a), Value::Integer(b)) => {
             let b = b as u32;
-            Ok(Value::Integer(a.pow(b) as i64))
+            Ok(Value::Integer(a.pow(b)))
         }
         (Value::Integer(a), Value::Float(b)) => Ok(Value::Float((a as f64).powf(b))),
         (Value::Float(a), Value::Float(b)) => Ok(Value::Float(a.powf(b))),
